Move scripts to their own subdirectories
