+++ src/celscript/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_subdirectory(legacy)
+if(ENABLE_CELX)
+  add_subdirectory(lua)
+endif()
