+++ src/celscript/lua/celx_category.h
