+++ src/celscript/lua/celx_misc.h
